refactor: change static dir/files to path(dirname) + fs to fs/promise + minor refactoring
