+++ src/fs/create.js
@@ -1,11 +1,17 @@
 import { writeFile, open, close } from 'fs';
+import path, { dirname } from 'path';
+import { fileURLToPath } from 'url';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+const ERR_CODE_EXIST = 'EEXIST';
+const ERR_MESSAGE = 'FS operation failed';
+const SUCCESS_MESSAGE = 'The file has been saved!';
+const FILE_NAME = path.join(__dirname, 'files', 'fresh.txt');
+const FILE_CONTENT = 'I am fresh and young';
 
 export const create = async () => {
-  const ERR_CODE_EXIST = 'EEXIST';
-  const ERR_MESSAGE = 'FS operation failed';
-  const SUCCESS_MESSAGE = 'The file has been saved!';
-  const FILE_NAME = 'src/fs/files/fresh.txt';
-  const FILE_CONTENT = 'I am fresh and young';
 
   open(FILE_NAME, 'wx', (err, fd) => {
     try {
