+++ src/fs/read.js
@@ -1,18 +1,21 @@
-import { readFile } from 'fs';
+import { readFile } from 'fs/promises';
+import path, { dirname } from 'path';
+import { fileURLToPath } from 'url';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 const ERR_MESSAGE = 'FS operation failed';
-const FILE_PATH = 'src/fs/files/fileToRead.txt';
+const FILE_PATH = path.join(__dirname, 'files', 'fileToRead.txt');
 
 export const read = async () => {
-  readFile(FILE_PATH, (err, data) => {
-    try {
-      if (err) throw new Error(ERR_MESSAGE);
+  try {
+    const content = await readFile(FILE_PATH);
 
-      console.log(data.toString('utf-8'));
-    } catch (e) {
-      console.error(e);
-    }
-  });
+    console.log(content.toString('utf-8'));
+  } catch (e) {
+    if (e) throw new Error(ERR_MESSAGE);
+  }
 }
