+++ src/fs/rename.js
@@ -1,16 +1,19 @@
-import fs from 'fs'
+import fs from 'fs/promises'
+import path, { dirname } from 'path';
+import { fileURLToPath } from 'url';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 const ERR_MESSAGE = 'FS operation failed';
-const FILE_NAME = 'src/fs/files/wrongFilename.txt';
-const NEW_FILE_NAME = 'src/fs/files/properFilename.md';
+const FILE_NAME = path.join(__dirname, 'files', 'wrongFilename.txt');
+const NEW_FILE_NAME = path.join(__dirname, 'files', 'properFilename.md');
 
 export const rename = async () => {
   try {
-    fs.rename(FILE_NAME, NEW_FILE_NAME, (err) => {
-      if (err) throw new Error(ERR_MESSAGE);
-    })
+    await fs.rename(FILE_NAME, NEW_FILE_NAME);
   } catch (e) {
-    console.error(e);
+    if (e) throw new Error(ERR_MESSAGE);
   }
 };
